move counter code to separate file
